Fixed error and bug in student functions (backend)

+++ backend/dashboard-service/src/main/java/com/gym/dashboard_service/controller/StudentApi.java
@@ -4,15 +4,17 @@ import com.gym.dashboard_service.model.studentmodel.StudentModel;
 import com.gym.dashboard_service.model.studentmodel.UpdateStudentModel;
 import com.gym.dashboard_service.service.StudentService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.HashMap;
 import java.util.List;
 
 @RestController
 @RequestMapping("/dashboard")
 public class StudentApi {
-
+    @Autowired
     StudentService service;
 
     @GetMapping(value = "/student/data")
@@ -21,7 +23,8 @@ public class StudentApi {
         return ResponseEntity.ok(data);
     }
     @PostMapping(value="/student/inset")
-    public  ResponseEntity<?> addNewStudent(@RequestBody StudentModel student,@RequestHeader("user_id") long userId){
+    public  ResponseEntity<?> addNewStudent(@RequestBody StudentModel student){
+        long userId=12345;
         service.addStudent(student,userId);
         return ResponseEntity.ok("success");
     }
@@ -32,10 +35,14 @@ public class StudentApi {
         return ResponseEntity.ok("success");
     }
 
-    @DeleteMapping(value="/student/remove")
-    public ResponseEntity<?> removeStudent(@RequestBody long studentId){
-        StudentModel student=service.getStudent(studentId);
-        service.deleteStudent(student);
+    @DeleteMapping(value="/student/remove/{dashId}/{studentID}")
+    public ResponseEntity<?> removeStudent(@PathVariable String dashId,@PathVariable long studentID){
+        StudentModel student=service.getStudent(studentID);
+        if (student.getDashboardId().equals(dashId)){
+            service.deleteStudent(student);
+        }else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("not student this dashboard");
+        }
         return ResponseEntity.ok("success");
     }
 }
